Rename initialization script
'init' is more likely to be understood at a glance by developers as it's a widely
used term to describe initialization/config routines in high level code.

Perhaps this change makes the framework feel more generic and less personal, so
if that doesn't sit right, then feel free to ignore it.